move initialize_ray out of __init__.py
